- Completed conversion for validator class-based tests. Just need to get to 100% 1
