+++ packages/validator-ajv6/src/types.ts
@@ -1,6 +1,11 @@
+import { ErrorSchema, RJSFValidationError } from '@rjsf/utils';
+
 export interface CustomValidatorOptionsType {
   additionalMetaSchemas?: ReadonlyArray<object>;
   customFormats?: { [k: string]: string | RegExp | ((data: string) => boolean) };
 }
 
-
+export interface ValidateFormDataReturnType<T> {
+  errors: RJSFValidationError[];
+  errorSchema: ErrorSchema<T>;
+}
